+++ app/src/main/java/com/example/tonetuner_v2/audio/audioSources/AudioSource.kt
@@ -2,5 +2,5 @@ package com.example.tonetuner_v2.audio.audioSources
 
 
 interface AudioSource {
-    fun getAudio(bufferSize: Int): List<Double>
+    fun getAudio(bufferSize: Int): List<Float>
 }
